Edit Folder Contructure
