+++ src/components/challengedetail/PostWrite.js
@@ -3,9 +3,8 @@ import styled from "styled-components";
 import Dialog from "@material-ui/core/Dialog";
 
 import { useDispatch, useSelector } from "react-redux";
-import { actionCreator as imageActions } from "../redux/modules/image";
-import { actionCreator as postActions } from "../redux/modules/post";
-import { consoleLogger } from "../redux/configureStore";
+import { actionCreator as imageActions } from "../../redux/modules/image";
+import { actionCreator as postActions } from "../../redux/modules/post";
 
 const PostWrite = ({ challengeId }) => {
   const dispatch = useDispatch();
