+++ src/components/mypage/CompletedChallenge.js
@@ -1,44 +1,31 @@
 import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
-import { actionCreators as myInfo } from "../redux/modules/mypage";
+import { actionCreators as myInfo } from "../../redux/modules/mypage";
+import { changeForm } from "./ChallengesInProgress";
 
-const changeForm = (dates) => {
-  let _month = dates?.map((m) => {
-    if (m.split("-")[1][0] === "0") {
-      return m.split("-")[1][1];
-    } else {
-      return m.split("-")[1];
-    }
-  });
+function CompletedChallenge() {
+  const dispatch = useDispatch();
 
-  let _date = dates?.map((d) => {
-    if (d.split("-")[2][0] === "0") {
-      return d.split("-")[2][1];
-    } else {
-      return d.split("-")[2];
-    }
-  });
-
-  return { _month, _date };
-};
-
-function ChallengesInProgress(props) {
   useEffect(() => {
-    dispatch(myInfo.getMyInfoDB());
+    dispatch(myInfo.getEndDB());
   }, []);
 
-  const dispatch = useDispatch();
   const myChallengeList = useSelector(
     (state) => state.mypage.myInfo.challengeList
   );
+
   const my_info = useSelector((state) => state.mypage.myInfo);
 
   const start = myChallengeList?.map(
     (list) => list.challengeStartDate.split("T")[0]
   );
+  const end = myChallengeList?.map(
+    (list) => list.challengeEndDate.split("T")[0]
+  );
 
   const { _month: start_month, _date: start_date } = changeForm(start);
+  const { _month: end_month, _date: end_date } = changeForm(end);
 
   return (
     <>
@@ -52,19 +39,18 @@ function ChallengesInProgress(props) {
               <div>
                 <div>
                   <span>
-                    {start_month[idx]}월 {start_date[idx]}일 부터 시작했어요!
+                    {start_month[idx]}월 {start_date[idx]}일 부터{" "}
+                    {end_month[idx]}월 {end_date[idx]}
+                    일까지 열심히 달성했어요!
                   </span>
+                  <span>완료!</span>
                 </div>
                 <div>
                   <span>{list.challengeTitle}</span>
                 </div>
                 <div>
                   <div>
-                    {/* <img
-                      src={}
-                      alt="profile_list"
-                      style={{ width: "2em", height: "2em" }}
-                    /> */}
+                    <img alt="" />
                   </div>
                   <div>
                     {list.participateSize && list.participateSize > 1 ? (
@@ -86,6 +72,7 @@ function ChallengesInProgress(props) {
     </>
   );
 }
+
 const ChallengeContent = styled.div`
   width: 100%;
   height: 100px;
@@ -103,6 +90,4 @@ const ChallengeImg = styled.div`
   }
 `;
 
-export { changeForm };
-
-export default ChallengesInProgress;
+export default CompletedChallenge;
